--- src/components/pages/home/homepage-data.ts
@@ -1,175 +1,0 @@
-const homepageData = {
-  jumbotron: {
-    id: 'jumbotron',
-    title: 'Deploy Ghost to AWS using RDS and EC2',
-    byline: 'featured course',
-    description: `Set up a Linux server on AWS EC2 with a relational database (AWS RDS) ready to run Ghost or other full stack JavaScript applications.`,
-    image:
-      'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/437/276/full/EGH_ghost-aws_1000.png',
-    background:
-      'https://res.cloudinary.com/dg3gyk0gu/image/upload/v1618861248/egghead-next-pages/deploy-ghost-to-aws-using-rds-and-ec2/background-deploy-ghost-to-aws-using-rds-and-ec2_2x.png',
-    path: '/courses/deploy-ghost-to-aws-using-rds-and-ec2-a3487caa',
-    slug: 'deploy-ghost-to-aws-using-rds-and-ec2-a3487caa',
-    instructor: {
-      name: 'Sam Julien',
-      slug: 'sam-julien',
-      path: '/q/resources-by-sam-julien',
-      twitter: 'samjulien',
-      image:
-        'https://d2eip9sf3oo6c2.cloudfront.net/instructors/avatars/000/000/324/medium/new_headshot_wedding.jpg',
-    },
-  },
-  'modern-layouts-with-css-grid': {
-    id: 'modern-layouts-with-css-grid',
-    title: 'Build Modern Layouts with CSS Grid',
-    name: 'Write modern CSS',
-    byline: 'Hiroko Nishimura・22m・Course',
-    description: `CSS Grid is a new technology that allows you to problem solve at the page-layout level. Learn how to create two-dimensional layouts that adapt content to the available space and intelligently size and position items within those bounds. `,
-    image:
-      'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/418/653/full/EGH_modern-layouts-css__1000.png',
-    background:
-      'https://res.cloudinary.com/dg3gyk0gu/image/upload/v1615913074/egghead-next-pages/build-modern-layouts-with-css-grid/background-build-modern-layouts-with-css-grid.svg',
-    path: '/courses/build-modern-layouts-with-css-grid-d3f5',
-    slug: 'build-modern-layouts-with-css-grid-d3f5',
-    instructor: {
-      name: 'Hiroko Nishimura',
-      slug: 'hiro-nishimura',
-      path: '/q/resources-by-hiro-nishimura',
-      twitter: 'hirokonishimura',
-      image:
-        'https://d2eip9sf3oo6c2.cloudfront.net/instructors/avatars/000/000/458/medium/IMG_20190627_100655_466.jpg',
-    },
-  },
-  tailwind: {
-    id: 'tailwind',
-    name: 'Write Modern CSS',
-    byline: 'Adam Wathan・3h 4m・Course',
-    description: `In this multi-module course, Adam Wathan - the creator of Tailwind - will teach you how to style your applications with modern, sustainable patterns. Modules include: `,
-    path: '/playlists/design-with-tailwind-css-masterclass-f0db',
-    slug: 'design-with-tailwind-css-masterclass-f0db',
-    title: 'Design with Tailwind CSS Masterclass',
-    image:
-      'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/348/912/full/tailwinds_masterclass_main_424_2x.png',
-    resources: [
-      {
-        title: 'Introduction to Tailwind and the Utility First Workflow',
-        path: '/playlists/introduction-to-tailwind-and-the-utility-first-workflow-ac67',
-      },
-      {
-        title: 'Design and Implement Common Tailwind Components',
-        path: '/playlists/design-and-implement-common-tailwind-components-8fbb9b19',
-      },
-      {
-        title: 'Build a Responsive Navbar with Tailwind',
-        path: '/playlists/build-a-responsive-navbar-with-tailwind-4d328a35',
-      },
-      {
-        title: 'Build and Style a Dropdown in Tailwind',
-        path: '/playlists/build-and-style-a-dropdown-in-tailwind-7f34fead',
-      },
-    ],
-  },
-  topics: {
-    id: 'topics',
-    name: 'Popular Topics',
-    resources: [
-      {
-        title: 'React',
-        path: '/q/react',
-        slug: 'react',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/026/thumb/react.png',
-      },
-      {
-        title: 'JavaScript',
-        path: '/q/javascript',
-        slug: 'javascript',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/205/thumb/javascriptlang.png',
-      },
-      {
-        title: 'CSS',
-        path: '/q/css',
-        slug: 'css',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/175/square_480/csslang.png',
-      },
-      {
-        title: 'Angular',
-        path: '/q/angular',
-        slug: 'angular',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/300/thumb/angular2.png',
-      },
-      {
-        title: 'Node',
-        path: '/q/node',
-        slug: 'node',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/256/thumb/nodejslogo.png',
-      },
-      {
-        title: 'TypeScript',
-        path: '/q/typescript',
-        slug: 'typescript',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/377/thumb/typescriptlang.png',
-      },
-      {
-        title: 'GraphQL',
-        path: '/q/graphql',
-        slug: 'graphql',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/001/034/thumb/graphqllogo.png',
-      },
-      {
-        title: 'AWS',
-        path: '/q/aws',
-        slug: 'aws',
-        image:
-          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/001/090/thumb/aws.png',
-      },
-
-      // {
-      //   title: 'Redux',
-      //   path: '/q/redux',
-      //   image:
-      //     'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/386/thumb/redux.png',
-      // },
-    ],
-  },
-  swag: {
-    id: 'swag',
-    name: 'swag',
-    title: '',
-    path: 'https://store.egghead.io',
-    resources: [
-      {
-        title: 'GraphQL Query Language Poster',
-        path: 'https://store.egghead.io/product/poster-graphql-query-language',
-        image:
-          'https://res.cloudinary.com/dg3gyk0gu/image/upload/v1606472163/next.egghead.io/store/poster-graphql-query-language.png',
-      },
-      {
-        title: 'egghead Crewneck',
-        path: 'https://store.egghead.io/product/egghead-crewneck',
-        image:
-          'https://res.cloudinary.com/dg3gyk0gu/image/upload/v1606474619/next.egghead.io/store/egghead-crewneck.png',
-      },
-      // {
-      //   title: 'Shirt',
-      //   path: 'https://store.egghead.io/product/egghead-shirt',
-      //   image:
-      //     'https://res.cloudinary.com/dg3gyk0gu/image/upload/v1606473230/next.egghead.io/store/egghead-shirt.jpg',
-      // },
-      // {
-      //   title: 'Beanie',
-      //   path: 'https://store.egghead.io/product/knit-beanie',
-      //   image:
-      //     'https://res.cloudinary.com/dg3gyk0gu/image/upload/v1606472232/next.egghead.io/store/knit-beanie.jpg',
-      // },
-    ],
-  },
-}
-
-export default homepageData
