--- src/components/pages/home/jumbotron/jumbotron.stories.mdx
@@ -1,37 +1,0 @@
-import {Meta, Story, Canvas} from '@storybook/addon-docs/blocks'
-import FeaturedResource from './'
-import {find} from 'lodash'
-import {sanityClient} from 'utils/sanity-client'
-
-export const resource = {
-  id: 'header',
-  title: 'Introduction to Cloudflare Workers',
-  byline: 'new course',
-  description:
-    "Follow along with Kristian Freeman as you build a localization engine that renders data based on the Edge location nearest to the application's user.",
-  image:
-    'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/418/892/full/EGH_IntroCloudFlareWorkers_Final.png',
-  path: '/playlists/introduction-to-cloudflare-workers-5aa3',
-  slug: 'introduction-to-cloudflare-workers-5aa3',
-  instructor: {
-    name: 'Kristian Freeman',
-    slug: 'kristian-freeman',
-    path: '/q/resources-by-kristian-freeman',
-    twitter: 'signalnerve',
-    image:
-      'https://d2eip9sf3oo6c2.cloudfront.net/instructors/avatars/000/000/469/square_128/kristian.jpeg',
-  },
-}
-
-<Meta title="Components/Home/Jumbotron" />
-
-# Featured Resource
-
-<Canvas>
-  <Story name="Default">
-    <div className="max-w-screen-xl mx-auto bg-gray-50 p-5">
-      <FeaturedResource resource={resource} />
-      <Primary />
-    </div>
-  </Story>
-</Canvas>
