+++ src/next-seo.json
@@ -24,7 +24,7 @@
     },
     "images": [
       {
-        "url": "https://res.cloudinary.com/dg3gyk0gu/image/upload/v1606837864/next.egghead.io/cards/default-card_2x.png"
+        "url": "https://res.cloudinary.com/dg3gyk0gu/image/upload/v1637345011/egghead-next-pages/home-page/root-og_2x.png"
       }
     ]
   }
